+++ GaussianElimination/GaussianElimination.Lib/Algorithms/SequentialAlgorithm.cs
@@ -5,14 +5,14 @@ namespace GaussianElimination.Lib.Algorithms;
 
 public class SequentialAlgorithm : IAlgorithm
 {
-    public float[] Solve(Matrix coefficients, float[] values)
+    public double[] Solve(Matrix coefficients, double[] values)
     {
         ForwardElimination(coefficients, values);
         BackwardSubstitution(coefficients, values);
         return values;
     }
 
-    private void ForwardElimination(Matrix coefficients, float[] values)
+    private void ForwardElimination(Matrix coefficients, double[] values)
     {
         int n = values.Length;
         for (int k = 0; k < n - 1; k++)
@@ -26,8 +26,8 @@ public class SequentialAlgorithm : IAlgorithm
 
             for (int i = k + 1; i < n; i++)
             {
-                float scale = coefficients[i, k] / coefficients[k, k];
-                coefficients.SubtractFromRow(i, k, scale, k+1);
+                double scale = coefficients[i, k] / coefficients[k, k];
+                coefficients.SubtractFromRow(i, k, scale, k + 1);
 
                 values[i] -= scale * values[k];
                 coefficients[i, k] = 0;
@@ -35,17 +35,18 @@ public class SequentialAlgorithm : IAlgorithm
         }
     }
 
-    // process back substitution on a upper triangle matrix
-    private void BackwardSubstitution(Matrix coefficients, float[] values)
+    // process back substitution on an upper triangle matrix
+    private void BackwardSubstitution(Matrix coefficients, double[] values)
     {
         int n = values.Length;
         for (int i = n - 1; i >= 0; i--)
         {
-            float sum = 0;
-            for (int j = i+1; j < n; j++)
+            double sum = 0;
+            for (int j = i + 1; j < n; j++)
             {
                 sum += coefficients[i, j] * values[j];
             }
+
             values[i] = (values[i] - sum) / coefficients[i, i];
         }
     }
